Fix fp16 type error.

+++ ggml.h
@@ -241,7 +241,7 @@ extern "C" {
 
 #ifdef __ARM_NEON
     // we use the built-in 16-bit float type
-    typedef __fp16 ggml_fp16_t;
+    typedef uint16_t ggml_fp16_t;
 #else
     typedef uint16_t ggml_fp16_t;
 #endif
